# packages/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manifest",
    "version": "0.1.0",
    "description": "Manifest's public API (§22). Generated from the control plane's route definitions; do not edit. Resource routes are under /v1 (D23.8); a breaking change is a new prefix beside it. Response objects may gain fields under /v1 — a client ignores fields it does not know. Every mutation carries an Idempotency-Key, and a session-bearing mutation carries Origin (§20)."
  },
  "servers": [
    {
      "url": "https://console.manifest.internal",
      "description": "The laptop platform (§21). The console and the API share this origin."
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
...
  "paths": {
...
    "/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "tags": [
          "projects"
        ],
        "summary": "The projects I am a member of",
        "description": "Every project the caller owns or collaborates on, newest first — for administrators too. The fleet is GET /v1/fleet.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The caller’s projects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the D23.7 envelope. This operation can answer: INTERNAL, REQUEST_INVALID, UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-manifest-error-codes": [
          "INTERNAL",
          "REQUEST_INVALID",
          "UNAUTHENTICATED"
        ]
      }
    },
...
    "/v1/projects/{projectId}/spec": {
      "get": {
        "operationId": "getSpec",
        "tags": [
          "projects"
        ],
        "summary": "The project’s newest valid manifest",
        "description": "manifest.yaml as last validated (§7). An invalid newest manifest answers SPEC_INVALID with its errors.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The spec.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Spec"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the D23.7 envelope. This operation can answer: INTERNAL, NOT_FOUND, REQUEST_INVALID, SPEC_INVALID, SPEC_NOT_FOUND, UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-manifest-error-codes": [
          "INTERNAL",
          "NOT_FOUND",
          "REQUEST_INVALID",
          "SPEC_INVALID",
          "SPEC_NOT_FOUND",
          "UNAUTHENTICATED"
        ]
      },
      "post": {
        "operationId": "validateSpec",
        "tags": [
          "projects"
        ],
        "summary": "Validate manifest.yaml at a commit",
        "description": "§22 step 3: reads manifest.yaml at the commit (HEAD by default), validates it (§7) and records the result. A sensitive diff (D9) is reported, not yet enforced.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "D23.6. One per user action, reused across retries of THAT action.",
            "schema": {
              "type": "string",
              "minLength": 8
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSpecRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The validation, valid or not — an invalid manifest is a recorded answer, not a refusal.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecValidation"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the D23.7 envelope. This operation can answer: AI_BACKEND_UNAVAILABLE, AI_CATALOGUE_EMPTY, CSRF_ORIGIN_REFUSED, FORBIDDEN, IDEMPOTENCY_KEY_REQUIRED, IDEMPOTENCY_KEY_REUSED, INTERNAL, NOT_FOUND, REQUEST_BODY_TOO_LARGE, REQUEST_INVALID, REQUEST_MEDIA_TYPE_UNSUPPORTED, SOURCE_GIT_FAILED, UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-manifest-error-codes": [
          "AI_BACKEND_UNAVAILABLE",
          "AI_CATALOGUE_EMPTY",
          "CSRF_ORIGIN_REFUSED",
          "FORBIDDEN",
          "IDEMPOTENCY_KEY_REQUIRED",
          "IDEMPOTENCY_KEY_REUSED",
          "INTERNAL",
          "NOT_FOUND",
          "REQUEST_BODY_TOO_LARGE",
          "REQUEST_INVALID",
          "REQUEST_MEDIA_TYPE_UNSUPPORTED",
          "SOURCE_GIT_FAILED",
          "UNAUTHENTICATED"
        ]
      }
    },
    "/v1/slugs/{slug}": {
      "get": {
        "operationId": "checkSlug",
        "tags": [
          "projects"
        ],
        "summary": "Would this project name work?",
        "description": "§23: answers exactly what project creation will, so a client can tell a person while they type. Always 200 — the answer is about the name, and a 4xx would make \"taken\" indistinguishable from \"not allowed to ask\". Says nothing about a holder. 60 a minute per person.",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The verdict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlugCheck"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the D23.7 envelope. This operation can answer: INTERNAL, RATE_LIMITED, REQUEST_INVALID, UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-manifest-error-codes": [
          "INTERNAL",
          "RATE_LIMITED",
          "REQUEST_INVALID",
          "UNAUTHENTICATED"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "manifest_session",
        "description": "Set by the CWL sign-in at /auth/login (§9). Delegated tokens arrive in P5b."
      }
    },
    "schemas": {
...
      "Audience": {
        "type": "object",
        "properties": {
          "scale": {
            "type": "string",
            "enum": [
              "solo",
              "class",
              "large_course",
              "public"
            ]
          },
          "burst": {
            "type": "string",
            "enum": [
              "steady",
              "synchronised"
            ]
          },
          "justification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "setBy": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "setAt": {
            "description": "An instant, ISO 8601 in UTC.",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "scale",
          "burst",
          "justification",
          "setBy",
          "setAt"
        ],
        "additionalProperties": false
      },
...
      "Environment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "kind": {
            "type": "string",
            "enum": [
              "sandbox",
              "staging",
              "production"
            ]
          },
          "hostname": {
            "description": "§23: `<slug>.<zone for this kind>`. Permanent.",
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "instance": {
            "description": "The instance the hostname reaches (§6 Route); for an app deployed before P4c, its newest instance. Null before any deploy.",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Instance"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "projectId",
          "kind",
          "hostname",
          "url",
          "instance"
        ],
        "additionalProperties": false
      },
...
      "ErrorCode": {
        "description": "Every code the API answers with (api/error-codes.ts). Stable: a client switches on it (§20).",
        "type": "string",
        "enum": [
          "AI_BACKEND_UNAVAILABLE",
          "AI_CATALOGUE_DISABLED",
          "AI_CATALOGUE_EMPTY",
          "AI_KEY_EXPIRED",
          "AI_KEY_REVOKED",
          "AI_MODEL_NOT_PERMITTED",
          "AI_MODEL_UNKNOWN",
          "AI_PROJECT_BUDGET_EXCEEDED",
          "AI_ROUTE_NOT_PERMITTED",
          "AI_UNMAPPED",
          "AI_USER_BUDGET_EXCEEDED",
          "BLUEPRINT_NOT_FOUND",
          "BUILD_INVALID_INPUT",
          "BUILD_LOG_INVALID_QUERY",
          "CONFIG_BUILD_CREDENTIAL_SECRET_REQUIRED",
          "CONFIG_CONTROL_PLANE_ORIGIN_PORT_MISMATCH",
          "CONFIG_INVALID",
          "CONFIG_LITELLM_MASTER_KEY_REQUIRED",
          "CONFIG_MASTER_SECRET_REQUIRED",
          "CSRF_ORIGIN_REFUSED",
          "DEPLOY_INVALID_INPUT",
          "EVENTS_UPGRADE_REQUIRED",
          "FORBIDDEN",
          "IDEMPOTENCY_KEY_REQUIRED",
          "IDEMPOTENCY_KEY_REUSED",
          "INTERNAL",
          "MEMBER_USER_NOT_FOUND",
          "NOT_FOUND",
          "PROJECT_INVALID_INPUT",
          "RATE_LIMITED",
          "RELEASE_AI_BUDGET_MISSING",
          "RELEASE_AI_DISABLED",
          "RELEASE_BLUEPRINT_NOT_FOUND",
          "RELEASE_BUILD_NOT_DEPLOYABLE",
          "RELEASE_BUILD_NOT_FOUND",
          "RELEASE_DIGEST_MISSING",
          "RELEASE_ENVIRONMENT_NOT_FOUND",
          "RELEASE_IMAGE_REPOSITORY_MISSING",
          "RELEASE_INVALID_INPUT",
          "RELEASE_LOCAL_IMAGE_ON_REMOTE_DRIVER",
          "RELEASE_MODEL_CLASSIFICATION_TOO_LOW",
          "RELEASE_MODEL_NOT_IN_CATALOGUE",
          "RELEASE_MODEL_UNCLASSIFIED",
          "RELEASE_NOT_FOUND",
          "RELEASE_PRODUCTION_GATE_UNAVAILABLE",
          "RELEASE_PROJECT_NOT_FOUND",
          "REQUEST_BODY_TOO_LARGE",
          "REQUEST_INVALID",
          "REQUEST_MEDIA_TYPE_UNSUPPORTED",
          "ROUTE_NOT_FOUND",
          "SAML_ASSERTION_REJECTED",
          "SAML_LOGIN_NOT_BOUND",
          "SAML_NO_PUID",
          "SAML_USER_UPSERT_FAILED",
          "SLUG_INVALID",
          "SLUG_RESERVED",
          "SLUG_TAKEN",
          "SOURCE_FOREIGN_REPO",
          "SOURCE_GIT_FAILED",
          "SOURCE_INVALID_SLUG",
          "SOURCE_PATH_ESCAPE",
          "SPEC_INVALID",
          "SPEC_NOT_FOUND",
          "UNAUTHENTICATED"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "description": "For a person. Never parse it; switch on `code`.",
                "type": "string"
              },
              "hint": {
                "description": "What to do about it.",
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ManifestError"
                }
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Instance": {
        "description": "A running (or once-running) copy of a release in one environment (§11). Never its driver or handle.",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "environmentId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "releaseId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "kind": {
            "type": "string",
            "enum": [
              "web",
              "worker",
              "cron"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "pending",
              "building",
              "provisioning",
              "starting",
              "healthy",
              "failed",
              "hibernated",
              "waking",
              "destroying",
              "gone"
            ]
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "description": "An instant, ISO 8601 in UTC.",
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "environmentId",
          "releaseId",
          "kind",
          "state",
          "lastSeenAt"
        ],
        "additionalProperties": false
      },
      "ManifestError": {
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ManifestErrorCode"
          },
          "path": {
            "description": "Where in manifest.yaml, dotted: `services.0.type`.",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "path",
          "message"
        ],
        "additionalProperties": false
      },
      "ManifestErrorCode": {
        "description": "A code inside `details`: §7 schema, §7 policy, or §25 blueprint compatibility.",
        "type": "string",
        "enum": [
          "BLUEPRINT_AI_UNSUPPORTED",
          "BLUEPRINT_AUTH_UNSUPPORTED",
          "BLUEPRINT_SCHEMA_VERSION_UNSUPPORTED",
          "BLUEPRINT_SERVICE_UNSUPPORTED",
          "SPEC_AI_BUDGET_REQUIRED",
          "SPEC_AI_DISABLED",
          "SPEC_ATTRIBUTE_NOT_REGISTERED",
          "SPEC_ATTRIBUTE_NOT_WHITELISTED",
          "SPEC_BUILD_BLOCK_FORBIDDEN",
          "SPEC_ENV_NAME_RESERVED",
          "SPEC_INVALID_BLUEPRINT_REF",
          "SPEC_INVALID_SLUG",
          "SPEC_INVALID_VALUE",
          "SPEC_MODEL_CLASSIFICATION_TOO_LOW",
          "SPEC_MODEL_UNCLASSIFIED",
          "SPEC_MODEL_UNKNOWN",
          "SPEC_NAME_SLUG_MISMATCH",
          "SPEC_PATH_EXPECTED",
          "SPEC_QUOTA_EXCEEDED",
          "SPEC_RESERVED_BLOCK_NOT_EMPTY",
          "SPEC_SERVICE_TYPE_UNKNOWN",
          "SPEC_UNKNOWN_KEY",
          "SPEC_YAML_PARSE_FAILED"
        ]
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "slug": {
            "description": "The project’s name, and the first label of every hostname it has (§23).",
            "type": "string"
          },
          "blueprint": {
            "description": "`name@major` (§25).",
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserSummary"
          },
          "audience": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Audience"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "description": "An instant, ISO 8601 in UTC.",
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "environments": {
            "description": "Present with `?expand=environments` (D23.1).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Environment"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "blueprint",
          "owner",
          "audience",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "ProjectList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Project"
        }
      },
      "SensitiveDiff": {
        "description": "D9. Reported, not yet enforced (P6).",
        "type": "object",
        "properties": {
          "sensitive": {
            "type": "boolean"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "sensitive",
          "fields"
        ],
        "additionalProperties": false
      },
      "SlugCheck": {
        "description": "§23: exactly what project creation will answer — advisory, since creation checks again.",
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "reasons": {
            "description": "Present when `available` is false: every reason that applies.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "enum": [
                    "SLUG_INVALID",
                    "SLUG_RESERVED",
                    "SLUG_TAKEN"
                  ]
                },
                "message": {
                  "type": "string"
                },
                "hint": {
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message",
                "hint"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "slug",
          "available"
        ],
        "additionalProperties": false
      },
      "Spec": {
        "type": "object",
        "properties": {
          "appSpecId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "commitSha": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$"
          },
          "spec": {
            "description": "manifest.yaml v1 as parsed and validated (§7). Its own JSON Schema is not published in P5a.",
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "appSpecId",
          "commitSha",
          "spec"
        ],
        "additionalProperties": false
      },
      "SpecValidation": {
        "type": "object",
        "properties": {
          "appSpecId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "commitSha": {
            "type": "string",
            "pattern": "^[0-9a-f]{40}$"
          },
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManifestError"
            }
          },
          "sensitiveDiff": {
            "$ref": "#/components/schemas/SensitiveDiff"
          }
        },
        "required": [
          "appSpecId",
          "commitSha",
          "valid",
          "errors",
          "sensitiveDiff"
        ],
        "additionalProperties": false
      },
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000)$"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ],
        "additionalProperties": false
      },
      "ValidateSpecRequest": {
        "type": "object",
        "properties": {
          "commitSha": {
            "description": "Defaults to the repository’s HEAD.",
            "type": "string",
            "pattern": "^[0-9a-f]{40}$"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}